requirements.txt
numpy==1.19.5
matplotlib==3.4.1
pandas==1.2.4
scikit-learn==0.24.1
keras==2.4.3
tensorflow==2.3.0
python-binance
TA-Lib
websocket_client